Mopho 2: Declared Identifier property_ControlType. In InfoForExposedParameters, added a controlType property to each parameter and declared and defined controlTypeFor(). Renamed the 'controls' filter to 'exposedParamControls'. Created 2_comp_ExposedParamControl.h and .cpp and declared and defined ExposedParamControl, adding instructions for building and deleting the three KnobAndAttachment classes I've  defined so far. In the KnobAndAttachment classes, renamed attachToExposedParameter() to attachKnobToExposedParameter() and renamed deleteAttachmentBeforeKnobToPreventLeaking() to deleteAttachmentBeforeKnobToPreventMemLeak().

+++ MophoGUI/MophoGUI 2.0/Source/exposedParamControls/0_comp_TooltipUpdaterForExposedParamControl.cpp
@@ -1,4 +1,4 @@
-#include "component_TooltipUpdaterForExposedParamControl.h"
+#include "0_comp_TooltipUpdaterForExposedParamControl.h"
 
 #include "../constants/constants_Identifiers.h"
 #include "../exposedParameters/ep_singleton_InfoForExposedParameters.h"
@@ -38,7 +38,7 @@ String TooltipUpdaterForExposedParamControl::generateTooltipText() {
         auto paramID{ info.IDfor(paramIndex) };
         auto paramaterPtr{ exposedParams->getParameter(paramID) };
         auto currentChoice{ roundToInt(paramaterPtr->convertFrom0to1(paramaterPtr->getValue())) };
-        auto choiceName{ info.verboseChoiceNameFor(currentChoice, paramIndex) };
+        auto choiceName{ info.verboseChoiceNameFor((uint8)currentChoice, paramIndex) };
         tooltipText += "Current Setting: " + choiceName;
     }
     return tooltipText;
